10.13.2014 Update
So I read through all the code and I have questions:

1. Where are the buttons being read in? Because whenever I try to add a
function all the buttons stop working? I also tried to just copy and
paste a second timer and the buttons still didn’t work. I tried to
write a countUp function and the buttons wouldn’t work either.

2. Going forward, do I have to create a new timer.js for every
instance?

3. Will each screen need its own javascript page?

I was able to make a restart button and right now I have is restarting
to 3:00. I tried to have it read in the #mins and #seconds but it never
worked since they aren’t global variables. I thought if I just passed
in updateTimer the original timer would read it but it just appears as
blank

Have a good night!