So close to being done!
I was able to get the home button and the back button working. I was
also able to update the prep time function.  Here are the problems I am
still having

1. How do I make the restart button disabled until stop is clicked?
Right now the whole thing goes haywire if the restart button is pressed
while the timer is going .

2. The activate the timer on the speech page you need to click a button
because it is looking for formats and there are none for speech. Do i
have to write a separate speech timers since there are no prep times
for formats?

3. How can I hide the home button and the back button on #home without
them disappearing from all other screens?